+++ NEWS.md
@@ -4,6 +4,7 @@
 * Argument `seq_names` in `asv2otu` function allow to clusterize sequences from a character vector of DNA.
 * Add a `blast_pq` function to blast the sequences of the `ref_seq` slot against a custom database
 * Add a `filter_asv_blast` function to filter ASV in phyloseq dataset using blast against a custom database
+* Add a `subset_taxa_pq` function to filter ASV based on a named conditional vector. Used in `filter_asv_blast`.
 
 # MiscMetabar 0.3